+++ src/main/kotlin/com/kstuca/shumovych/master/library/repository/BookRepository.kt
@@ -7,7 +7,7 @@ import org.springframework.data.repository.PagingAndSortingRepository
 import org.springframework.stereotype.Repository
 
 @Repository
-interface BookRepository : PagingAndSortingRepository<BookModel, Int>{
+interface BookRepository : PagingAndSortingRepository<BookModel, Long>{
 
     @Query(value = "SELECT new com.kstuca.shumovych.master.library.domain.GenreCount(genre, COUNT(*) as count) FROM BookModel GROUP BY genre", nativeQuery = false)
     fun findAllGenresWithCount(): List<GenreCount>
